added CCSV_to_file

+++ counters.h
@@ -2,14 +2,15 @@
 extern "C" {
 #endif
 
+#include <stdint.h>
 
 #ifndef CCSV_COUNTERS_H
 #define CCSV_COUNTERS_H
 
 struct CCSV_Counters {
-    unsigned strings,
-             max_columns,
-             chars;
+    size_t strings,
+           max_columns,
+           chars;
 };
 
 void CCSV_Counters_init(struct CCSV_Counters *const counters);
